查询基金会详情	queryFund	完成	2017/2/23 查询合约的追踪详情	queryContractTrack	完成	2017/2/23	合约的爱心数量，合约的余额情况，已筹集金额 捐款或提款的交易详情	queryTransDetail	完成	2017/2/23	捐款完成 发起提款，提款记录查询	queryDrawDetail	完成	2017/2/23

+++ src/main/java/com/ecloudtime/model/User.java
@@ -11,7 +11,7 @@ public class User {
     private String addr;
     private String type;
     private String balance;
-    private int total; //我已捐赠的款
+    private long total; //我已捐赠的款
     private String totalStr;
     private List<DonorContribution> contributions ;
 
@@ -69,11 +69,11 @@ public class User {
 		this.addr = addr;
 	}
 
-	public int getTotal() {
+	public long getTotal() {
 		return total;
 	}
 
-	public void setTotal(int total) {
+	public void setTotal(long total) {
 		this.total = total;
 	}
